Review fix #3
